Rename to namanager (#15)
* Rename to namanager

* Rename to namanager

* Fix error: no such file or directory and test not trigger error
