+++ README.md
@@ -1,5 +1,5 @@
-# File Checker
-[![Build Status](https://travis-ci.org/iattempt/FileChecker.svg?branch=master)](https://travis-ci.org/iattempt/FileChecker) [![codecov](https://codecov.io/gh/iattempt/FileChecker/branch/master/graph/badge.svg)](https://codecov.io/gh/iattempt/FileChecker)
+# namanager
+[![Build Status](https://travis-ci.org/iattempt/namanager.svg?branch=master)](https://travis-ci.org/iattempt/namanager) [![codecov](https://codecov.io/gh/iattempt/namanager/branch/master/graph/badge.svg)](https://codecov.io/gh/iattempt/namanager)
 
 A checker/detector could check name of file/directory is expectable or not.
