--- file_checker/__init__.py
@@ -1,5 +1,0 @@
-from file_checker.main import import_settings, main
-
-__all__ = [
-    'import_settings', 'main'
-]
