+++ namanager/__init__.py
@@ -1,0 +1,5 @@
+from namanager.main import import_settings, main
+
+__all__ = [
+    'import_settings', 'main'
+]
