+++ namanager/enums.py
