+++ namanager/settings.json
