+++ namanager/util.py
