+++ setup.py
@@ -1,7 +1,7 @@
 import setuptools
 import os
 
-NAME = 'File Checker'
+NAME = 'namanager'
 VERSION = '0.0.1'
 DESCRIPTION = (
     'A checker could check name of file/directory is expectable or not.'
@@ -15,9 +15,9 @@ AUTHOR = (
     'Arnav Borborah (arnavb@github)'
 )
 AUTHOR_EMAIL = 'iattempt.net@gmail.com'
-URL = 'https://github.com/iattempt/FileChecker'
-PACKAGES = {'file_checker'}
-KEYWORDS = ['checker', 'formatter', 'filename']
+URL = 'https://github.com/iattempt/namanager'
+PACKAGES = {'namanager'}
+KEYWORDS = ['name', 'manager', 'checker', 'formatter', 'filename']
 PYTHON_REQUIRES = '==3.4.*,==3.5.*,==3.6.*'
 
 setuptools.setup(
