+++ tests/test_main.py
@@ -1,10 +1,10 @@
-from file_checker.main import import_settings, main
+from namanager.main import import_settings, main
 
 
 class TestMain():
     def test_import_settings(self):
-        import_settings(settings_file='file_checker/settings.json')
+        import_settings(settings_file='namanager/settings.json')
 
     def test_main(self):
-        import_settings(settings_file='file_checker/settings.json')
+        import_settings(settings_file='namanager/settings.json')
         main()
